+++ CHANGELOG.md
@@ -15,6 +15,10 @@ incremented for features.
 
 * cli: fix dns in NODE_OPTIONS ([#928](https://github.com/project-serum/anchor/pull/928)).
 
+### Breaking
+
+* spl: Move permissioned markets into dex repository ([#962](https://github.com/project-serum/anchor/pull/962)).
+
 ## [0.18.0] - 2021-10-24
 
 ### Features